refactor to rename task directory to tasks
