+++ src/main/java/Ozymandias/Tasks/Task.java
@@ -1,4 +1,4 @@
-package Ozymandias.Task;
+package Ozymandias.Tasks;
 
 public abstract class Task {
     private static int nextId = 1;          // Auto-increment ID if you want unique IDs
