+++ src/main/java/Ozymandias/ui/TaskList.java
@@ -2,7 +2,7 @@ package Ozymandias.ui;
 
 import java.util.LinkedHashMap;
 
-import Ozymandias.Task.Task;
+import Ozymandias.Tasks.Task;
 
 public class TaskList {
     private LinkedHashMap<Integer, Task> tasks;
