+++ src/test/java/Ozymandias/Tasks/EventsTest.java
@@ -1,8 +1,6 @@
-package Ozymandias.Task;
+package Ozymandias.Tasks;
 
 import org.junit.jupiter.api.Test;
-
-import Ozymandias.Task.Events;
 
 public class EventsTest {
